requirements.txt
requests
beautifulsoup4
python-dotenv
Flask
Flask-SocketIO
eventlet
